No comment

+++ file_manager/lib/download_files.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:io';
+import 'dart:ui';
 
 import 'package:archive/archive_io.dart';
 import 'package:file_manager/delete_files.dart';
@@ -9,28 +10,33 @@ import 'package:path_provider/path_provider.dart';
 import 'package:share_plus/share_plus.dart';
 import 'package:toast/toast_message.dart';
 
+enum DownloadResult { success, dismissed, unavailable, noFiles, error, copyDownloadFolderError }
+
 /// Encodes `filesToEncode` into a ZIP folder and shares it
-Future<bool> shareFiles(List<File> filesToEncode, bool deleteOriginal, DateTime date, [String? userName]) async {
+Future<DownloadResult> shareFiles(List<File> filesToEncode, {required bool deleteOriginal, required DateTime date, String? userName, Rect? sharePositionOrigin}) async {
   if (filesToEncode.isEmpty) {
     print('No files were passed to share');
-    return false;
+    return DownloadResult.noFiles;
   }
   if (filesToEncode.length == 1) {
-    bool result = true;
+    DownloadResult result = DownloadResult.success;
     try {
       // share file
-      final ShareResult shareResult = await Share.shareXFiles([XFile(filesToEncode.first.path)]);
-      if (shareResult.status == ShareResultStatus.dismissed) result = false;
-      if (shareResult.status == ShareResultStatus.unavailable) result = false;
+      final ShareResult shareResult = await Share.shareXFiles([XFile(filesToEncode.first.path)], sharePositionOrigin: sharePositionOrigin);
+      if (shareResult.status == ShareResultStatus.dismissed) result = DownloadResult.dismissed;
+      if (shareResult.status == ShareResultStatus.unavailable) result = DownloadResult.unavailable;
       // delete all the files
-      await Future.delayed(const Duration(seconds: 2));
-      if (deleteOriginal) for (final File file in filesToEncode) await file.delete();
-    } catch (_) {
-      result = false;
+      if (result == DownloadResult.success) {
+        await Future.delayed(const Duration(seconds: 2));
+        if (deleteOriginal) for (final File file in filesToEncode) await file.delete();
+      }
+    } catch (e) {
+      print(e);
+      result = DownloadResult.error;
     }
     return result;
   }
-  bool result = true;
+  DownloadResult result = DownloadResult.success;
   try {
     // compress files
     late String zipPath;
@@ -39,17 +45,18 @@ Future<bool> shareFiles(List<File> filesToEncode, bool deleteOriginal, DateTime 
     zipPath = userName.isUsername ? '$dirPath/${dateString}_Measurement_$userName.zip' : '$dirPath/${dateString}_Measurement.zip';
     final ZipFileEncoder encoder = ZipFileEncoder()..create(zipPath);
     for (final File file in filesToEncode) await encoder.addFile(file);
-    encoder.close();
+    await encoder.close();
     // share compressed zip file
-    final ShareResult shareResult = await Share.shareXFiles([XFile(zipPath)]);
-    if (shareResult.status == ShareResultStatus.dismissed) result = false;
-    if (shareResult.status == ShareResultStatus.unavailable) result = false;
+    final ShareResult shareResult = await Share.shareXFiles([XFile(zipPath)], sharePositionOrigin: sharePositionOrigin);
+    if (shareResult.status == ShareResultStatus.dismissed) result = DownloadResult.dismissed;
+    if (shareResult.status == ShareResultStatus.unavailable) result = DownloadResult.unavailable;
     // delete all the files
     await Future.delayed(const Duration(seconds: 2));
     if (deleteOriginal) for (final File file in filesToEncode) await file.delete();
     await File(zipPath).delete();
-  } catch (_) {
-    result = false;
+  } catch (e) {
+    print(e);
+    result = DownloadResult.error;
   }
   return result;
 }
@@ -64,11 +71,17 @@ Future<bool> shareFiles(List<File> filesToEncode, bool deleteOriginal, DateTime 
 /// `date` is used to name the ZIP folder.
 ///
 /// Returns `true` if the operation was successful, `false` otherwise. May throw an exception, that is however overridden in release mode.
-Future<bool> downloadFiles(List<File> filesToEncode, [bool deleteOriginal = false, DateTime? date, String? userName]) async {
+Future<DownloadResult> downloadFiles(List<File> filesToEncode, {bool deleteOriginal = false, DateTime? date, String? userName, Rect? sharePositionOrigin}) async {
   // Unable to save to Downloads folder on iOS
-  if (Platform.isIOS) return shareFiles(filesToEncode, deleteOriginal, date ?? DateTime.now());
+  if (Platform.isIOS)
+    return shareFiles(
+      filesToEncode,
+      deleteOriginal: deleteOriginal,
+      date: date ?? DateTime.now(),
+      sharePositionOrigin: sharePositionOrigin,
+    );
   // else (on Android) download to Downloads folder
-  bool result = true;
+  DownloadResult result = DownloadResult.success;
   date ??= DateTime.now();
 
   /// Get directory paths
@@ -77,24 +90,34 @@ Future<bool> downloadFiles(List<File> filesToEncode, [bool deleteOriginal = fals
   if (dirPath == null) {
     // if no Downloads folder found, use sharing
     print('No Downloads folder found');
-    return shareFiles(filesToEncode, deleteOriginal, date);
+    return shareFiles(
+      filesToEncode,
+      deleteOriginal: deleteOriginal,
+      date: date,
+      sharePositionOrigin: sharePositionOrigin,
+    );
   }
 
   /// Do copy or zip
   if (filesToEncode.isEmpty) {
     // No files were provided
     print('No files were passed to download');
-    return false;
+    return DownloadResult.noFiles;
   }
   if (filesToEncode.length == 1) {
     // Copy one file to Download folder
     final String fileName = filesToEncode.first.path.fileNameWithExtension;
-    result = await _copy(filesToEncode.first, File('$dirPath/$fileName'));
-    if (result) {
+    result = await _copy(filesToEncode.first, File('$dirPath/$fileName')) ? DownloadResult.success : DownloadResult.copyDownloadFolderError;
+    if (result == DownloadResult.success) {
       unawaited(MyToast.showToast(message: 'Downloaded $fileName to Downloads folder'));
       await Future.delayed(const Duration(milliseconds: 500));
     } else {
-      result = await shareFiles(filesToEncode, deleteOriginal, date);
+      result = await shareFiles(
+        filesToEncode,
+        deleteOriginal: deleteOriginal,
+        date: date,
+        sharePositionOrigin: sharePositionOrigin,
+      );
     }
     if (deleteOriginal) await deleteFiles(filesToEncode);
     return result;
@@ -105,12 +128,17 @@ Future<bool> downloadFiles(List<File> filesToEncode, [bool deleteOriginal = fals
     final String zipName = userName.isUsername ? '${dateString}_Measurement_$userName.zip' : '${dateString}_Measurement.zip';
     final String zipPath = '$tempDir/$zipName';
     await _compressFiles(zipPath, filesToEncode);
-    result = await _copy(File(zipPath), File('$dirPath/$zipName'));
-    if (result) {
+    result = await _copy(File(zipPath), File('$dirPath/$zipName')) ? DownloadResult.success : DownloadResult.copyDownloadFolderError;
+    if (result == DownloadResult.success) {
       unawaited(MyToast.showToast(message: 'Downloaded $zipName to Downloads folder'));
       await Future.delayed(const Duration(milliseconds: 500));
     } else {
-      result = await shareFiles(filesToEncode, deleteOriginal, date);
+      result = await shareFiles(
+        filesToEncode,
+        deleteOriginal: deleteOriginal,
+        date: date,
+        sharePositionOrigin: sharePositionOrigin,
+      );
     }
     await Future.delayed(const Duration(milliseconds: 300));
     if (deleteOriginal) await deleteFiles(filesToEncode);
@@ -118,7 +146,7 @@ Future<bool> downloadFiles(List<File> filesToEncode, [bool deleteOriginal = fals
     return result;
   }
   print('If the code made it this far, the developer made some mistake');
-  return false;
+  return DownloadResult.error;
 }
 
 String? _getDownloadsFolder() {
@@ -146,6 +174,6 @@ Future<bool> _copy(File original, File copy) async {
 Future<File> _compressFiles(String path, List<File> files) async {
   final ZipFileEncoder encoder = ZipFileEncoder()..create(path);
   for (final File file in files) await encoder.addFile(file);
-  encoder.close();
+  await encoder.close();
   return File(path);
 }
